refactor(board): gemini-review-issues behoben in sticky-note-pad und board-toolbar
- DOM-Kopplung entfernt: board-boundary-check aus StickyNotePadComponent in
  BoardToolbarComponent verschoben (Trennung der Zuständigkeiten)
- NgZone.runOutsideAngular für mousemove-tracking, re-entry nur für Signal-Updates
- DOCUMENT-Token statt direktem document-Zugriff (SSR-kompatibel)
- totes AppModule-Export-Artefakt entfernt
- togglePinned: !this.pinned() statt ternary
- #darken: unterstützt nun auch 3-stellige Hex-Codes (#abc)
